pyproject.toml
[tool.poetry]
name = "my-stacking-framework"
version = "0.1.0"
description = ""
authors = ["Hassan Abedi <hassan.abedi.t@gmail.com>"]
readme = "README.md"
packages = [{ include = "my-stacking-framework" }]

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "^1.0.0"
scikit-learn = "^1.3.2"
scipy = "^1.11.4"
pandas = "^2.1.3"
matplotlib = "^3.8.2"
lightgbm = "^4.1.0"
optuna = "^3.4.0"
catboost = "^1.2.2"
jax = "^0.4.20"
keras-core = "^0.1.7"
jaxlib = "^0.4.20"
skrebate = "^0.62"
sklearn-relief = "^1.0.0b2"
xgboost = "^2.0.2"
polars = "^0.19.19"
pyarrow = "^14.0.1"
tpot = "^0.12.1"
tabulate = "^0.9.0"
nbconvert = "^7.13.1"
setuptools = "^69.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
